Add files via upload
This was my favorite homework during my first semester of graduate study. It's nothing too crazy, just a few different cluster analysis techniques. However, it was a ton of fun to apply these ideas so such a unique dataset (the link to the Kaggle dataset can be found at the top of the code).